auth_service/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0

sqlalchemy[asyncio]==2.0.44
asyncpg==0.30.0
alembic==1.13.2

databases[asyncpg]==0.9.0

pydantic==2.12.4
pydantic-core==2.41.5
typing_extensions==4.15.0

passlib[argon2]==1.7.4

PyJWT==2.10.1
python-dotenv==1.0.1

cryptography==42.0.8
